requirements.txt
# =============================================================================
# PiUPS Breakout Garden IoT Sensor App - Dependencies
# Optimized for Raspberry Pi with 2GB RAM
# =============================================================================

# Pimoroni Breakout Garden I2C Sensors
bh1745>=0.0.4
ltr559>=0.0.6
vl53l1x>=0.0.5
bme680>=2.0.0
lsm303d>=0.0.4
smbus2>=0.4.0

# OLED Display (SH1106 128x128)
luma.oled>=3.8.0
luma.core>=2.4.0

# Snowpipe Streaming v2 REST API
requests>=2.28.0
PyJWT>=2.8.0
cryptography>=41.0.0

# System monitoring
psutil>=5.9.0

# Testing (install with: pip install -r requirements.txt[test])
pytest>=7.4.0
pytest-mock>=3.12.0
responses>=0.24.0
